Added ability to log the exe's request for nif files
The exe writes nif errors to resmgr.log without telling which file the error is in.
Enabling USE_NIF_LOGGING in globalDefineALT will now log the exe's nif file requests.
This way if there is an error, the previous line in the log should tell which entry contains the error.