Added AlertWindow to allow a consistent and configurable error popup window
Wrapper class for Windows' MessageBoxA. This can now be called using human readable code.
Configured to be placed on top even when the game is in full screen.
Vanilla's gDLL->MessageBox is a frontend to MessageBoxA, through not configurable.
AlertWindow allows multiple icons, multiple buttons and provides info on which button was pressed.