Fix #854: Buying deep water ships from king without any suitable colony no longer crashes the game
Triggering DIPLOEVENT_ROYAL_INTERVENTION will now place ships in Europe if no deep water port is present
in the new world. Fixed crash when trying to print the name of a NULL city.

Rewrote LocationFlags from enum to struct. This allows flags to be independent bools instead of one hot.
This is not only more readable code and in the debugger, it allows adding Europe as a fallback option.