Cleanup: OrderData is now using enum types instead of ints for everything
OrderTypes is set by constructor and can't be changed later.
Made the data private and added accessors functions. This avoids lots of typecasting.
Assert checking if data is written to as the same type as it is being read as later.
For instance unit() returns a UnitTypes reference and will assert fail if the order is a building.
Added unions to store the data types. While strictly not needed, it makes it readable in the debugger.

Note: the vanilla warning about data locations being fixed and causing issues if it changes might not be
true. It doesn't look like OrderData itself is transferred to the exe and instead it's split into
an argument for each variable. Still the memory layout is preserved just to be safe.