Fixed crash related to fishing boats changing profession (start/end fishing)
Problem was that CvSelectionGroup::continueMission calls pLoopUnit->gatherResource, which has the ability
to clear the mission queue. The queue is checked to not be empty before the call and assumes it to not be
empty afterwards without checking. This sometimes resulted in using a NULL mission pointer.

Whaling boats use the same code, so this is likely also a fix for those too, through it's not confirmed.