Removed Project Files from release zip files
The release zip building script will now ignore all the content of Project Files, leaving an empty dir.
This is because startup time in the exe seems to increase with the number of files in the mod, possibly
doing so exponentially. Removing the source files will result in reaching the main menu faster.
99%+ of the people downloading a stable release will never look at the source code anyway.

This means a release will consist of two zip files, one with a dll and one with the source.
GitHub will automatically create the latter and we have no way to prevent that anyway.