Fix #889: fixed crash due to using ::getUnit without checking return value for NULL
::getUnit has NULL as a valid return value, yet most of the code will use the returned unit pointer
without testing if it's NULL. This has resulted in crashes in the past and it did it again.

This commit has reviewed all 282 calls to ::getUnit and protected each from crashing due to NULL pointers.

Note: it can be argued that it's a bug if we get a NULL pointer, but we do and the end result are crashes.

Two calls have yet to be protected against NULL pointers. Both are in pathfinding.