#259 Names of cities, ships, generals and admirals are now TXT KEYs
This not only allow translation, but also fixes character encoding issues where colony names were
displayed incorrectly.

This also makes civilization info xml pure ASCII and will work with all tools without specifying encoding.
This is the only info xml file with encoding issues and as such we now have more xml tool freedom.

Note: the names are currently still in civilization infos. They are however not read and will be removed
later when it can easily be done so by scripting and/or other tools.