BUTTONPOPUP: each popup is now a dedicated class with no code anywhere else
Popup windows are called by the exe using 5 different functions. This means to add a new one, 5 functions
will need to be updated in addition to to the code for popup itself.

Now it's based on a base class with 5 virtual functions. This way none of the 5 access functions will
need updating in case a new popup is added.

The only code outside the class is a single line declaration to map enum to class<enum>.