Fixed infinite loop in CvSelectionGroup::canStartMission
Problem: the unit node moved to the next node at the end of the loop. A continue made it not move.

Solution:
New function CvSelectionGroup::getUnitNodeLoop fetches the unit pointer and move the node to the next.
Using this function it's no longer possible to do one and not the other, even if loop iterations
end prematurely.