#929: Started organizing CvDLLWidgetData.cpp
Introducing the concept of one class for each WidgetTypes (currently only for WIDGET_PEDIA_JUMP_TO_UNIT).
This allows everything related to that widget to be on the screen at once instead of spread across the
entire file (vanilla spread code for a single widget into up to 7 functions)

With a single init location for each widget, iData1+2 is set and typecasted consistently.