XML Editor: tags not configured in editor config xml will now have a gray background
It was supposed to be seen as just a string, but the editor window for that string is broken.
Rather than fixing it, the solution is to just not use it and force using a config for all tags.
A tag with no config is often an enum type to another file, meaning it's "missing config" anyway.