XMLReader: opened xml files are now stored in memory until CiXMLFileReader::clearCache() is called
Opening an xml file will now call a class, which stores the files in a vector.
Trying to open one will return the already opened and passed document file, which is useful due to
the multiple rounds of xml reading, which is now 3 as postLoadSetup has the xml file too.

This will also allow multiple xml files to share an already opened schema file.