Code bugfix: MIN_CHAR and MIN_SHORT are now always negative
Vanilla bug. The defines in CvGameCoreDLL.h results in losing the sign if viewed as an int.
Vanilla treats MIN_CHAR as 0x80 without a type, hence it becomes 128 if stored in an int.
Now all the min/max are const ints as this reveals the value when mouse over (cleaner code).
Signed awareness also prevents MAX_UNSIGNED_INT from being stored as a signed int by the compiler.