InfoArray: reduced memory usage
Each stored variable can now be stored as following:
- short: previously only option
- -128 to 127: standard byte
- -1 to 254: Ideal for storing xml indexes and NONE (-1)

All 3 modes can be used in dynamic and static mode (pointer and static array setup in a union)

The two new modes use a single byte as opposed to previously always 2 bytes.
This allows a static array of 4 in the union with the pointer.
Lots of xml data can now be stored statically (4 yields, 2units+2 professions etc)
As a result, many realistic use cases of InfoArray will no longer allocate memory outside of the class.

Note: storage mode is determined when writing to an InfoArray, hence runtime.
This means the same InfoArray in an xml info class might not use the same mode for all entries.
Instead the most optimal will be picked for each instance, including static/dynamic switching.