Bug fix for FF Logic + Colopedia Minor Update
I realized today that some Founding Father categories actually require LESS of their specific point for some reason. I reworked the code to ensure that if a FF requires any points besides Political, the highest of those will be the primary, regardless of how high the Political Points required is.