River Fishing Nets Improvement Added
Prior solution to Fishing Nets working on Large Rivers didn't work due to a misunderstanding on my part of how the game enabled Fishing Nets. This solution is the only solution I can think of at the moment, which is to add a new River Fishing Nets improvement. Doing this ended up being a pain, but it works.

I also had to modify CvGameTextMgr.cpp to stop saying that River Fishing Nets allowed land units to cross, as there was a generic check for TERRAIN_LARGE_RIVERS making an improvement valid that supposedly meant all land units could cross. Not all improvements build on Large Rivers allow crossing, so I had to make this function more specific for river ferries.