TBB cleanup: it's now possible to compile with multithreading turned off
Created an internal API for multithreading, which works in singlethreaded mode too.
This way code outside this API layer will generally not be concerned with number of threads.
As a result the same code will now be executed even if TBB isn't used.

Added a globally available bool function to determine if multithreading is currently in use.
This is used to assert check CvRandom to see if the thread-unsafe code is called multithreaded.