Fix #923: Citizens can now leave colonies as colonists even if they have no movement points left
If a unit has zero or negative moves left, then the AI profession finding code will create a NULL unit.
The reason is that the AI code assumes all units can switch to default profession with no fallback.

An example of how this can go wrong: scout->farmer->kicked out. Scout spent 90 movement before joining.
Next turn the unit is kicked out. 90 is too much for a colonist so leaving the city fails.
doTurn is called on all citizens later the same turn where 60 movement points are restored.
Units outside of cities are handled after cities so it's correct to not call doTurn before kicking.

The fix/workaround is to bypass the movement points left check if new profession is the default one.

It's unknown if this will completely fix #923, but it fixes the specific case I tested this on.