Fix CTD caused by pUnit being NULL in CvPlayer::buyYieldUnitFrom
When adding X units of a yield to a cargo hold that already has Y yield units, the added yield unit is effectively consumed if the sum is less than the limit of a cargo hold and thus the added unit is killed since it is merged into the existing yield unit.
We failed to handle this case which resulted in pUnit being NULL (i.e. killed)