Fix: Spawned PoWs would capture defender plot no matter if other defenders were present or not.
Previous code didn't check for the presence of other defenders on the defender tile before spawning a PoW on the defender's coordinates, resulting in cities full of defenders being taken in an instant.

New code now checks for presence of other defenders, and if there are any when a PoW is to be spawned, the PoW will instead spawn on the attacker's tile.

Tested in all instances, both inside and outside cities.

Also fixed a typo in a variable I noticed here in CvUnit.cpp.