Fix: CvPlot::m_workingCity can no longer point to a non-existing city or cities out of range
Vanilla bug

If a city is removed and there is no other city in the area to gain control of the plot, then vanilla
will keep the city ID in the plot. CvPlot::getWorkingCity() will still work as intended, so mostly fine.
The problem occurs if a new city happens to get precisely the same ID, which is possible now that it is
free. This will then cause issues with plot reporting the city as working city while the city can't
find it while looping plots within range.

Savegame correcting code included so savegames work even if this has gone wrong prior to saving.