Fix: Americas - Random maps no longer spawn all Europeans in Southern Ice
Previous code tried to randomize the starting plot selection algorithm after running the algorithm, but this function is completely misused due to passing a massive int resulting in an unsigned short overflow which resulted in wacky behavior. I've disabled this final 'randomization', since the current randomization algorithm for use for starting plot selection functions very well as is.