Update: FaireWeatherTweakEx no longer obsesses over hilly and mountainous terrain
My experience with the updated FaireWeather is that it is quite good. Further improvements can be made, but this is a definite improvement that reduces the previous obsession with hilly and mountainous terrain that FaireWeather had due to an undetected issue with how the script calculated the amount of tiles that would be hilly/mountainous.

I've also updated the terrain roughness TXT_KEYs to be more explicit in terms of percentages as to what the player is going to get, approximately. The percentages aren't perfect due to some preexisting erosion functions in the script.