Fix: After capture fixes, new issue related to civilian capture appeared that I've handled.
After fixing the PoW capturing, I began running into an interesting issue where captured civilians began spawning on an incorrect tile, sometimes even on enemy ships pushed out of the city.

I went on to fix the ship capture problem, and have finally gotten around to taking a look at the civilian issue.

Moving the PoW code down to the Advance code somehow screwed up the civilian code in some subtle way, or perhaps it's been present for a while, but now I've applied a fix that marks all capturable civilians as being captured, and then kills them in a separate loop.

Somehow the first civilian fought would be captured, then the second civilian would be captured on the correct tile, but every civilian after that would appear on an incorrect tile. But that issue is now fixed. Works outside cities too, and goods capture still works correctly as well.