Cleaned up CvActionInfo generation
Now new entries are stored directly in a vector rather than temp storage of 4 arrays and a vector.
AcActionInfos are stored in a vector rather than a vector of pointers (faster looping).
Prepared code for not assigning actions to everything like right now (we don't need 604 actions)
However all actions are still included.

Added ActionTypes class with functions for looping and similar.

CvHotkeyInfo and it's interaction with CvActionInfo still needs cleanup.