Civilians loaded in ships will no longer be captured when a civilian is fought in a city
I finally came to the realization as to what was causing captured civilian units to appear on ships outside a city. It turns out the code didn't check for whether or not a civilian was loaded into a transport before marking them for capture. Now it checks if they're in a transport, and if it's a water transport, they will not be captured. Land transport-mounted civvies still get captured though.