Added autogenerate.exe
This major change to the build process should not affect people, who just wants to compile.

Added a standalone C++ project called autogenerate, which over time will replace perl scripts.
Compiling WTP will compile autogenerate automatically and then execute it to get header files.

This being a C++ project will allow us to use better code organization as well as a debugger.
With clean C++ it will be easier to expand on the code generation to make precisely what we want.

Currently it can read xml and generate header files from it, through they are not used yet.
First goal is to add classes of enums, which are interchangable (can even replace) xml enums.
Being classes, this means we can add functions to them and while they are still "just an int", they
will be more productive to work with, like adding functions to get the relevant info class.

Other might come later, like a rework of EnumMap to make it more readable for humans and compiler.