Possible fix for issue #1116
It seems that the reason why emphasize yields doesn't work is because they
have been isolated from code.

This fix makes the enphasize yields to be added as a final calculation to yield value.

I've also changed the code that makes Cannon values to be set to 0 to be only if they are either Normal, or de-emphasized.