Autogenerate: new InfoArray classes now have function implementations
Currently they "only" support getLength, get(enums/int) and assigning from a vector.
The plan is to not implement support between EnumMap and InfoArray.
Instead a new class/function will take both as arguments when interaction is needed.
This split will likely make the implementation a lot easier to maintain.